# api-rest-gogin/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api de teste de modelagem para documentação de API´s no Swagger.",
        "title": "Exemplo para teste do Swagger",
        "termsOfService": "https://policies.google.com/terms?hl=pt-BR\u0026fg=1",
        "contact": {
            "name": "Suport Devs",
            "url": "http://www.exemplo.com",
            "email": "contato@example.com"
        },
        "license": {
            "name": "Licença: GPLv3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/index",
    "paths": {
        "/alunos": {
            "get": {
                "description": "Rota que recupera e exibe todos os alunos cadastrados no db.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Recupera e exibe alunos.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Rota que cria um novo aluno a partir de parâmetros do tipo objeto json e salva no db.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Cria um novo aluno.",
                "parameters": [
                    {
                        "description": "Modelo de Aluno",
                        "name": "aluno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AlunoPostReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/alunos/cpf/{CPF}": {
            "get": {
                "description": "Rota que recupera e exibe  um aluno a partir do CPF passado (URL/path).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Recupera e exibe aluno a partir do CPF.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CPF do Aluno",
                        "name": "CPF",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/alunos/rg/{RG}": {
            "get": {
                "description": "Rota que recupera e exibe  um aluno a partir do RG passado (URL/path).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Recupera e exibe aluno a partir do RG.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RG do Aluno",
                        "name": "RG",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/alunos/{ID}": {
            "get": {
                "description": "Rota que recupera e exibe  um aluno a partir do ID passado (URL/path).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Recupera e exibe aluno a partir do ID.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Aluno",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Rota que deleta um aluno do db a partir de seu ID passado através da URL.",
                "tags": [
                    "Alunos"
                ],
                "summary": "Deleta o aluno do db.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Aluno",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso!"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Rota que edita os dados de um aluno a partir de seu ID passado através da URL.",
                "tags": [
                    "Alunos"
                ],
                "summary": "Edita os dados do aluno.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Aluno",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Modelo de Aluno",
                        "name": "aluno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AlunoPostReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso!",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/index": {
            "get": {
                "description": "Rota que recupera a pagina principal com dados dos alunos cadastrados no sistemab.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Pagina principal de alunos.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "models.Aluno": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "rg": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.AlunoPostReq": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "rg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}